Updated: Fetch tab file on editor load
